{"schema":"decodex/commit/1","summary":"Split app bridge request and event parsing surfaces","authority":"manual"}
